Added some comments/TODOs.
Created beginning of an outline for what the program should run through so as to make a more concrete plan. Also added TODO: Create an algorithm that defines how "bomby" a card is.

"Bomby-ness" is defined on a numerical scale based on the quality of the card for a player's deck relative to other cards in a pack. Maybe something like where X/Y, where X is a card with a,b,c... traits and Y is the number of cards left in pack array.